// src/app/core/models/footer/footer.ts
export enum EnumEstiloDoFooter {
    Light = 'footer-light',
    Dark = 'footer-dark',
}

export enum EnumEnderecoNeo {
    GrupoNeo = 'Uma empresa do Grupo Neoenergia.',
    Coelba = 'Avenida Edgar Santos, 300, Cabula VI - Salvador/BA - CEP 41181-900 CNPJ: 15.139.629/0001-94',
    Cosern = 'Rua Mermoz, 150, Baldo - Natal/RN - CEP 59025-250 CNPJ: 08.324.196/0001-81',
    Celpe = 'Av. João de Barros, 111, Boa Vista - Recife/PE - CEP 50050-902 CNPJ: 10.835.932/0001-08',
    Elektro = 'Rua Ary Antenor de Souza, 321, Jardim Nova América - Campinas/SP - CEP 13053-024 CNPJ: 02.328.280/0001-97'
}